FOOD DATABASE QUERIES/requirements.txt
mysql-connector-python==9.2.0
mysqlclient==2.2.7
